Wrapping useSearchQuery with suspense
